% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kladd_Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
dijkstra(Graph, init_node)
}
\arguments{
\item{Graph}{A data frame}

\item{init_node}{An integer.}
}
\value{
A vector that contains the shortest path from one node to the others.
}
\description{
Function that calculates the shortest path from one node to the others.
}
\examples{
dijkstra(wiki_graph, 1)
}

