% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MinTestKod.R
\name{euclidian}
\alias{euclidian}
\title{Euclidian algorithm}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidian(e, t)
}
\arguments{
\item{e}{A number.}

\item{t}{A number.}
}
\value{
The largest number that divides both of them without leaving a remainder.
}
\description{
Function that takes two numbers and calcualtes the greatest common divisor.
}
\examples{
euclidian(100, 1000)
}

