\name{KevGus-package}
\alias{KevGus-package}
\alias{KevGus}
\docType{package}
\title{
\packageTitle{KevGus}
}
\description{
\packageDescription{KevGus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KevGus}
\packageIndices{KevGus}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KevGus}

Maintainer: \packageMaintainer{KevGus}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
