% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MinTestKod.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Data frame with nodes}
\format{A data frame with 18 rows and 3 columns:
\describe{
   \item{v1}{current node}
   \item{v2}{connecting node}
   \item{w}{distance between nodes}
   }}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
Dataset containing the different nodes for Djikstra function.
}
\keyword{datasets}

