% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MinTestKod.R
\name{euclidian}
\alias{euclidian}
\title{Take two numbers.}
\usage{
euclidian(e, t)
}
\arguments{
\item{e}{A number.}

\item{t}{A number.}
}
\value{
The largest number that divides both of them without leaving a remainder.
}
\description{
Take two numbers.
}
\examples{
euclidian(100, 1000)
}

